refactor: local import no longer needed after #1194
